# perf: add inlining for range/iter and benchmark